requirements.txt
Django==1.11
Pillow==5.4.1
boto3==1.9.104
botocore==1.12.104
certifi==2018.11.29
chardet==3.0.4
colorama==0.2.5
dj-database-url==0.5.0
django-forms-bootstrap==3.1.0
django-storages==1.7.1
docutils==0.14
html5lib==0.999
idna==2.8
jmespath==0.9.4
psycopg2==2.7.7
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.5.2
stripe==2.21.0
unattended-upgrades==0.1
urllib3==1.24.1
wheel==0.24.0
